+++ CHANGELOG.rst
@@ -15,8 +15,9 @@ Changelog
   renamed to ``named_coef_``. Accessing ``named_coefs_`` will raise a
   ``DeprecationWarning``.
 
-- :func:`~ivmodels.quadric.Quadric.__format__` now prints the same as
-  :func:`~ivmodels.confidence_set.ConfidenceSet.from_quadric.__format__` if the quadric
+- :class:`~ivmodels.quadric.Quadric`'s ':func:`~ivmodels.quadric.Quadric.__format__` now
+  prints the same as :class:`~ivmodels.confidence_set.ConfidenceSet`'s
+  :func:`~ivmodels.confidence_set.ConfidenceSet.__format__` if the quadric
   is one-dimensional.
 
 0.7.0 - 2025-06-03